Only render UserDetailsTooltip on web
