+++ src/components/UserDetailsTooltip/index.js
@@ -1,53 +1,17 @@
-import React, {useCallback} from 'react';
-import {View, Text} from 'react-native';
-import {withOnyx} from 'react-native-onyx';
-import Str from 'expensify-common/lib/str';
-import lodashGet from 'lodash/get';
-import _ from 'underscore';
-import Avatar from '../Avatar';
-import Tooltip from '../Tooltip';
-import {propTypes, defaultProps} from './userDetailsTooltipPropTypes';
-import styles from '../../styles/styles';
-import ONYXKEYS from '../../ONYXKEYS';
-import * as UserUtils from '../../libs/UserUtils';
+import PropTypes from 'prop-types';
 
-function UserDetailsTooltip(props) {
-    const userDetails = lodashGet(props.personalDetailsList, props.accountID, props.fallbackUserDetails);
-    const renderTooltipContent = useCallback(
-        () => (
-            <View style={[styles.alignItemsCenter, styles.ph2, styles.pv2]}>
-                <View style={styles.emptyAvatar}>
-                    <Avatar
-                        containerStyles={[styles.actionAvatar]}
-                        source={UserUtils.getAvatar(userDetails.avatar, userDetails.accountID)}
-                    />
-                </View>
+const propTypes = {
+    /** Children to wrap with Tooltip. */
+    children: PropTypes.node.isRequired,
+};
 
-                <Text style={[styles.mt2, styles.textMicroBold, styles.textReactionSenders, styles.textAlignCenter]}>
-                    {String(userDetails.displayName).trim() ? userDetails.displayName : ''}
-                </Text>
-
-                <Text style={[styles.textMicro, styles.fontColorReactionLabel]}>
-                    {String(userDetails.login || '').trim() && !_.isEqual(userDetails.login, userDetails.displayName) ? Str.removeSMSDomain(userDetails.login) : ''}
-                </Text>
-            </View>
-        ),
-        [userDetails.avatar, userDetails.displayName, userDetails.login, userDetails.accountID],
-    );
-
-    if (!userDetails.displayName && !userDetails.login) {
-        return props.children;
-    }
-
-    return <Tooltip renderTooltipContent={renderTooltipContent}>{props.children}</Tooltip>;
-}
+/**
+ * @param {propTypes} props
+ * @returns {ReactNodeLike}
+ */
+const UserDetailsTooltip = (props) => props.children;
 
 UserDetailsTooltip.propTypes = propTypes;
-UserDetailsTooltip.defaultProps = defaultProps;
-UserDetailsTooltip.displayName = 'UserDetailsTooltip';
+UserDetailsTooltip.displayName = 'Tooltip';
 
-export default withOnyx({
-    personalDetailsList: {
-        key: ONYXKEYS.PERSONAL_DETAILS_LIST,
-    },
-})(UserDetailsTooltip);
+export default UserDetailsTooltip;
